Move class PageCollection into Foundation namespace
