+++ packages/framework/src/RouteCollection.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Collection;
 /**
  * Pseudo-Router for Hyde.
  *
- * @see \Hyde\Framework\PageCollection
+ * @see \Hyde\Framework\Foundation\PageCollection
  * @see \Hyde\Framework\Testing\Feature\RouteTest
  *
  * This is not a router in the traditional sense that it decides where to go.
